tests/wpt/web-platform-tests/longtask-timing/longtask-in-sibling-iframe.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>LongTask Timing: long task in sibling iframe</title>
<body>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
  async_test(t => {
    window.addEventListener("message", t.step_func(e => {
      assert_equals(e.data, "longtask+same-origin+frame");
      t.done();
    }));
}, "Performance longtask entries in parent are observable in child iframe");
</script>

<iframe src="resources/subframe-observing-longtask.html"></iframe>

<script>
  /* Create an iframe that generates a long task. */
  var iframe = document.createElement('iframe');
  iframe.id = 'child-iframe-id';
  iframe.name = 'child-iframe-name';
  document.body.appendChild(iframe);
  iframe.src = 'resources/subframe-with-longtask.html'
</script>
</body>
